requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
stripe==7.8.0
django-filter==23.5
django-extensions==3.2.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
djoser==2.2.2
djangorestframework-simplejwt==5.3.0
django-cleanup==8.0.0